Updated Campaign  with new methods
I added the Date and appropriate formatter to display when a session is
created, and during each log. This will help players keep track of when.

I changed the log to be a 'StringBuilder' instead of plain 'String';
which will allow me to manipulate it easier then before.

I have added stubs for beginSession() and endSession().

I have added the appropriate JavaDoc information.